Fixed HighStounSelector
Fixed HighStounSelector.
Now this menu is displayed immediately after starting the game from the main menu, before loading the first level of the race.
