+++ Assets/Prefabs/UI/Builders/MainMenuUIBuilder.prefab
@@ -45,5 +45,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _viewsPrefab:
-  - {fileID: 735601388137290407, guid: 01cf66c11572c41439382735b69f564a, type: 3}
+  - {fileID: 735601388137290407, guid: ebd531b5100176e48b8dfd657da686e2, type: 3}
   - {fileID: 6762105179401652709, guid: 830bb3b460df145bfb46cc52d7eb4513, type: 3}
+  - {fileID: 2034671531895404175, guid: 5dcfe5038e4bf394db11456b3f4978fa, type: 3}
