--- Assets/Prefabs/UI/Popups/HighStoneChooseMenu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0387fc7e30907ce44af455d42de4e976
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
