+++ Assets/Prefabs/UI/Popups/HighStoneChooseMenu.prefab.meta
