--- Assets/Prefabs/UI/Popups/StartSplashScreen/StartSplashScreen.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 01cf66c11572c41439382735b69f564a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
